+++ src/main/java/org/kdb/inside/brains/view/console/ConsoleOptionsPanel.java
@@ -21,6 +21,7 @@ public class ConsoleOptionsPanel extends JPanel {
     private final JBCheckBox expandList = new JBCheckBox("Vector");
     private final JBCheckBox expandDict = new JBCheckBox("Dictionary");
     private final JBCheckBox expandFlip = new JBCheckBox("Table");
+    private final JBCheckBox consoleBackground = new JBCheckBox("Use an instance color for console background");
     private final JBIntSpinner floatPrecisionEditor = new JBIntSpinner(7, 0, ConsoleOptions.MAX_DECIMAL_PRECISION);
     private final ComboBox<ConsoleSplitType> splitTypes = new ComboBox<>(ConsoleSplitType.values());
 
@@ -41,6 +42,7 @@ public class ConsoleOptionsPanel extends JPanel {
         formBuilder.addLabeledComponent("Float precision: ", floatPrecisionEditor);
         createSplitTypes(formBuilder);
         addExpandPanel(formBuilder);
+        formBuilder.addComponent(consoleBackground);
 
         add(formBuilder.getPanel());
     }
@@ -90,6 +92,7 @@ public class ConsoleOptionsPanel extends JPanel {
         consoleOptions.setExpandList(expandList.isSelected());
         consoleOptions.setExpandDict(expandDict.isSelected());
         consoleOptions.setExpandTable(expandFlip.isSelected());
+        consoleOptions.setConsoleBackground(consoleBackground.isSelected());
         return consoleOptions;
     }
 
@@ -107,5 +110,6 @@ public class ConsoleOptionsPanel extends JPanel {
         expandList.setSelected(consoleOptions.isExpandList());
         expandDict.setSelected(consoleOptions.isExpandDict());
         expandFlip.setSelected(consoleOptions.isExpandTable());
+        consoleBackground.setSelected(consoleOptions.isConsoleBackground());
     }
 }